Refactor engine layout into modular hierarchy
