+++ src/engine/nn/nnue.cpp
@@ -1,7 +1,7 @@
 #include "nnue.h"
 
 #include "nnue_loader.h"
-#include "../eval.h"
+#include "../eval/eval.h"
 
 #include <cstdint>
 #include <utility>
