+++ src/engine/search/thread.c
