+++ src/engine/util/util.c
